refactor playerHand
