+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { default as CityLines } from './CityLines';
 export { default as ResearchStation } from './ResearchStation';
-export { default as Card } from './Card';
+export { default as PlayerHand } from './PlayerHand';
 export { default as Board } from './Board';
 export { default as ActionFooter } from './ActionFooter';
 export { default as GameHeader } from './GameHeader';
